# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Simplify ERP API",
        "description": "A sessão abaixo contém a documentação da API utilizada para o Simplify ERP.",
        "contact": {
            "name": "Felipe Silva",
            "email": "felipe.allware@gmail.com"
        },
        "license": {
            "name": "Licença MIT",
            "url": "https://github.com/felipesilva15/simplify-erp-api/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local"
        },
        {
            "url": "https://sandbox.simplify-erp.felipesilva15.com.br",
            "description": "Sandbox"
        },
        {
            "url": "https://api.simplify-erp.felipesilva15.com.br",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/api/partner/partner-types/{id}/activity-logs": {
            "get": {
                "tags": [
                    "PartnerType"
                ],
                "summary": "List activity logs of a partner type",
                "operationId": "9728477260e19e5809a450795e46c365",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Partner type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Partner type activity logs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ActivityLogCollection"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/partner/partner-types": {
            "get": {
                "tags": [
                    "PartnerType"
                ],
                "summary": "List all partner types",
                "operationId": "a5676a76eaf761f47222b7e9dfab2b23",
                "parameters": [
                    {
                        "name": "filters[id][eq]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filters[name][like]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filters[created_at][gte]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filters[updated_at][lte]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/sortsParam"
                    },
                    {
                        "$ref": "#/components/parameters/perPageParam"
                    },
                    {
                        "$ref": "#/components/parameters/pageParam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PartnerType list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/PartnerTypeCollection"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "PartnerType"
                ],
                "summary": "Registers a partner type",
                "operationId": "4ddab14bbbba0663b08eb7381b0dc85f",
                "requestBody": {
                    "description": "Data for creating a new partner type",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePartnerTypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered partner type data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/PartnerTypeResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/partner/partner-types/{partner_type}/edit": {
            "get": {
                "tags": [
                    "PartnerType"
                ],
                "summary": "Get data to edit a partner type",
                "operationId": "0065448e3da9b3c2ce076fcb8fb91cb7",
                "parameters": [
                    {
                        "name": "partner_type",
                        "in": "path",
                        "description": "PartnerType ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PartnerType data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/PartnerTypeResource"
                                                },
                                                "warnings": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Este recurso não pode ser editado."
                                                    },
                                                    "nullable": true
                                                },
                                                "meta": {
                                                    "properties": {
                                                        "editable": {
                                                            "type": "boolean",
                                                            "example": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/partner/partner-types/lookup": {
            "get": {
                "tags": [
                    "PartnerType"
                ],
                "summary": "Lookup partner types",
                "operationId": "e8eb67b50d46ef6ee4439a6d74d999dc",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Query for avaiable fields",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "keys[]",
                        "in": "query",
                        "description": "Keys to find",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Partner type lookup list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/PartnerTypeLookupCollection"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "required": [
                    "success",
                    "message",
                    "data"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operação realizada com sucesso"
                    },
                    "data": {
                        "nullable": true
                    },
                    "warnings": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "links": {
                        "type": "object",
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "sortsParam": {
                "name": "sorts",
                "in": "query",
                "description": "Fields for sorting separated by commas. Use '-' to sort descending",
                "required": false,
                "schema": {
                    "type": "string"
                },
                "example": "-id"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Insira o token JWT",
                "name": "token",
                "in": "cookie"
            }
        }
    },
...
}